Refactor: Ensure programs are installed per-user
Changes the installation of `easyeffects` and `neovim` to install them for specific users instead of system-wide. This prevents unnecessary installation for all users and aligns with typical user-specific application configurations.

Additionally, this commit refactors the cloning of Git repositories to use `system.userActivationScripts` instead of `system.activationScripts`. This is a more idiomatic NixOS approach for user-specific tasks and simplifies the activation process.