Refactor: Remove unnecessary polkit and power ops configuration
This commit removes the explicit configuration for Polkit and user power operations.
The `GDK_BACKEND` environment variable is simplified to only use `wayland` as it's the default and more performant on Wayland sessions.
Additionally, Seahorse and Libvirt are no longer enabled by default, as they are less commonly required for a base desktop setup. This simplifies the configuration and reduces potential conflicts or unnecessary services running.