feat: Add zathura and update libreoffice to still
Adds zathura for PDF viewing and switches the libreoffice package to `libreoffice-still` for better stability.

Also, a small change was made to the `Refresh.sh` script to redirect `waybar` output to `/dev/null` to prevent potential clutter in logs. Finally, `application/yaml` is now associated with `nvim-wrapper.desktop` to improve YAML file handling.